Moved prediction to classification file.
