+++ aroma/tests/test_utils.py
@@ -1,20 +1,8 @@
 """Tests for aroma.utils."""
-import os
-
 import numpy as np
 import pandas as pd
 import pytest
 from aroma import utils
-
-
-def test_classification(classification_overview):
-    """Test aroma.utils.classification and ensure classifications come out the same."""
-    clf_overview_df = pd.read_table(classification_overview)
-    test_df = clf_overview_df[["edge_fract", "csf_fract", "max_RP_corr", "HFC"]]
-    test_df, metadata = utils.classification(test_df, {})
-    true_classifications = clf_overview_df["classification"].tolist()
-    test_classifications = test_df["classification"].tolist()
-    assert true_classifications == test_classifications
 
 
 def test_load_motpars_manual(motion_parameters):
